+++ packages/uno-serverless-cli/test/unit/samples/entities/user.ts
@@ -17,4 +17,5 @@ export interface User {
     id: string,
     name?: string;
   }>;
+  related: User;
 }