+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - CLI: generate-schemas: added -config option to specify tsconfig.json file (Fix #60)
 - HttpClient: Added debug mode (Fix #62)
 - Added clear() and list() to KeyValueRepository (Fix #57)
+- Added ForwardEventDispatcher (Fix #63)
 
 ## [0.49.0] - 2018-08-23
 ### Added