+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.50.0] - 2018-09-05
 ### Added
 - SQSEventPublisher to publish events using AWS Simple Queue Service
 - SQSEvent & SQSEventBatch as specialized handlers for Simple Queue Service-backed lambdas